[ASTGen] Split ASTBridging.cpp into multiple files
ASTBridging.cpp was getting too big.
