+++ lib/AST/Bridging/CMakeLists.txt
@@ -1,0 +1,14 @@
+target_sources(swiftAST PRIVATE
+  ASTContextBridging.cpp
+  DeclAttributeBridging.cpp
+  DeclBridging.cpp
+  DeclContextBridging.cpp
+  DiagnosticsBridging.cpp
+  ExprBridging.cpp
+  GenericsBridging.cpp
+  MiscBridging.cpp
+  PatternBridging.cpp
+  PluginBridging.cpp
+  StmtBridging.cpp
+  TypeAttributeBridging.cpp
+  TypeReprBridging.cpp)
